Move octokit function to a separate file
