+++ packages/github-client/src/index.ts
@@ -1,9 +1,9 @@
-import { createAppAuth, createOAuthUserAuth } from "@octokit/auth-app";
-import { Octokit } from "@octokit/core";
+import { createAppAuth } from "@octokit/auth-app";
 import { Client, cacheExchange, fetchExchange } from "urql";
 import { graphql as gql } from "./graphql";
 import type { GitHubAuthConfig } from "./types";
 export * from "./types";
+export * from "./octokit";
 
 export const IssueNodeIdQuery = gql(/* GraphQL */ `
   query IssueNodeIdQuery($name: String!, $owner: String!, $issueNumber: Int!) {
@@ -68,38 +68,4 @@ export async function graphql(authConfig: GitHubAuthConfig) {
 			headers: { authorization: `Bearer ${token}` },
 		},
 	});
-}
-
-export function octokit(authConfig: GitHubAuthConfig) {
-	switch (authConfig.strategy) {
-		case "github-installation": {
-			return new Octokit({
-				authStrategy: createAppAuth,
-				auth: {
-					appId: authConfig.appId,
-					privateKey: authConfig.privateKey,
-					installationId: authConfig.installationId,
-				},
-			});
-		}
-		case "github-token": {
-			return new Octokit({ auth: authConfig.token });
-		}
-		case "github-app-user": {
-			return new Octokit({
-				authStrategy: createOAuthUserAuth,
-				auth: {
-					clientId: authConfig.clientId,
-					clientSecret: authConfig.clientSecret,
-					clientType: "oauth-app",
-					token: authConfig.token,
-					refreshToken: authConfig.refreshToken,
-				},
-			});
-		}
-		default: {
-			const _exhaustiveCheck: never = authConfig;
-			throw new Error(`Unhandled authConfig strategy: ${_exhaustiveCheck}`);
-		}
-	}
 }
